+++ src/main/java/com/ryszard/domain/vo/Department.java
@@ -4,23 +4,23 @@ import java.util.Objects;
 
 public class Department {
 
-    private Long departmentId;
+    private int departmentId;
 
     private String departmentName;
 
     public Department() {
     }
 
-    public Department(Long departmentId, String departmentName) {
+    public Department(int departmentId, String departmentName) {
         this.departmentId = departmentId;
         this.departmentName = departmentName;
     }
 
-    public Long getDepartmentId() {
+    public int getDepartmentId() {
         return departmentId;
     }
 
-    public void setDepartmentId(Long departmentId) {
+    public void setDepartmentId(int departmentId) {
         this.departmentId = departmentId;
     }
